week_12/day_03/primarche/Primarche/src/main/java/main/Product.java
package main;

public abstract class Product {

    private double buyPrice;
    private double sellPrice;
    private String manufacturer;
    private long id;
    private String productName;

    public Product(double buyPrice, double sellPrice, String manufacturer, long id, String productName) {
        this.buyPrice = buyPrice;
        this.sellPrice = sellPrice;
        this.manufacturer = manufacturer;
        this.id = id;
        this.productName = productName;
    }

    public double getBuyPrice() {
        return buyPrice;
    }

    public void setBuyPrice(double buyPrice) {
        this.buyPrice = buyPrice;
    }

    public double getSellPrice() {
        return sellPrice;
    }

    public void setSellPrice(double sellPrice) {
        this.sellPrice = sellPrice;
    }

    public String getManufacturer() {
        return manufacturer;
    }

    public void setManufacturer(String manufacturer) {
        this.manufacturer = manufacturer;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getProductName() {
        return productName;
    }

    public void setProductName(String productName) {
        this.productName = productName;
    }

    public double calculateMarkup(){
        return this.getSellPrice() - this.getBuyPrice();
    }
}
